Bai6.cpp

+++ 28techdrive/Bai6.cpp
@@ -6,6 +6,6 @@ using namespace std;
 
 int main(){
   int x1,y1,x2,y2; cin >> x1>>y1>>x2>>y2;
-  float khoangcach = sqrt(pow(x2-x1,2)+pow(y2-y1,2));
+  double khoangcach = sqrt(pow(x2-x1,2)+pow(y2-y1,2));
   cout << fixed << setprecision(2) << khoangcach << endl;
 }
